moved test_join_import tests into test_importer as JoinImporterTest
